Autohinting makes modern fonts worse most of the time
https://wiki.archlinux.org/index.php/Font_configuration#Autohinter

The autohinter attempts to do automatic hinting and disregards any
existing hinting information. Originally it was the default because
TrueType2 fonts were patent-protected but now that these patents have
expired there is very little reason to use it. It does work better with
fonts that have broken or no hinting information but it will be strongly
sub-optimal for fonts with good hinting information. Generally common
fonts are of the later kind so autohinter will not be useful. Autohinter
is disabled by default.