Halt after installation is complete
Default behavior was to reboot, but that may lead to infinite install loop
if USB stick is selected as the default boot device in BIOS

Poweroff could have been as useful, but then I will have to manually reach
into the motherboard to boot again. With halt I can just cycle the power.