Just poweroff after installation is complete
Halting was not a good idea. It was done in an attempt to make use of
'Action on AC power restore: Boot', but going that way meant there is no
indication when headless installation is complete, therefore there is no
way to know when it's ok to cycle the power.

Full poweroff instead gives a clear indication when it's ok to remove the
installation media and to power the machine back on.