Drop all connections from and to malicious hosts
If a host was caught doing malicious stuff (currently only ssh bruteforce
is detected) it will be banned from all incoming and outgoing connections
for 5 minutes.

Rate limit should always be on port 22 even if there is no sshd or if sshd
is listening on a non-default port, because failed connection attempts
also count against the limit.