Install nftables ahead of time
There is a short period of time between establishing Wireguard VPN and
configuring nftables on outer firewall when inner edge loses connectivity
with WAN. That is exactly the time we would try to install nftables. And
it would fail, obviously.

Now nftables package is installed ahead of its role to avoid such
transient errors