Do not test docker-compose role here
The service being deployed (microblog-server) will not launch successfully
without a valid Telegram token. Providing a token for a test instance and
maintaining it inside repo secrets is too much for a simple playbook
check.

Now we assume that the role will work correctly because it's been tested
separately, and testing the container itself is out of scope for now.