Explicitly notify about successful termination of command
`go build` does not output anything when it finishes successfully, and
under watchexec it's difficult to distinguish if a command just hangs
there or if it has exited with 0.

Status lines are only printed for non-zero codes, so now we add an
ephemeral 'OK' after every command - it will be erased by the next
iteration.