bash: Do not record tmux parent shell session
Recording Tmux parent terminal is messy and redundant: child terminals
will produce a much better scrollback file.

We can not easily tell `script` to stop recording so we just delete the
file as soon as we detect that the recording is not needed. This does not
stop the recording: writing to deleted file still happens, but that data
will be garbage collected upon process exit eventually.